checks: make deps

+++ misc/latest.sh
@@ -20,7 +20,15 @@ target_path=
 for path in "$@" /usr/local/bin /usr/bin ~/.local/bin 'C:\Program Files\Git\usr\bin'; do
     case :"$path": in
         *:"$path":*)
-            mkdir -p "$path" >/dev/null 2>&1 || true
+            if ! mkdir -p "$path" >/dev/null 2>&1; then
+                continue
+            fi
+            monkey_test="$path"/monkey
+            if [ -f "$monkey_test" ] && [ -w "$monkey_test" ]; then
+                target_path="$path"
+                echo "Selected target path: $target_path"
+                break
+            fi
             if touch "$path"/monkey >/dev/null 2>&1; then
                 rm   "$path"/monkey >/dev/null 2>&1
                 target_path="$path"
@@ -32,14 +40,14 @@ done
 if [ -z "$target_path" ]; then
     fatal "Could not find a suitable target path among $PATH"
 fi
+target="$target_path"/monkey
 
 echo Looking for latest tag of $slug
-latest_tag_url=$(curl --silent --location --output /dev/null --write-out '%{url_effective}' $page/releases/latest)
+latest_tag_url=$(curl -# --fail --location --output /dev/null --write-out '%{url_effective}' $page/releases/latest)
 latest_tag=$(basename "$latest_tag_url")
 latest_version=monkey/"$latest_tag"
-echo Latest tag: "$latest_tag"
 
-if [ "$(monkey --version 2>&1 || true)" = "$latest_version" ]; then
+if [ "$("$target" --version 2>&1 || true)" = "$latest_version" ]; then
     echo "Latest tag is already installed. You're good to go"
     exit 0
 fi
@@ -49,10 +57,11 @@ case "$exe" in
     CYGWIN*|MINGW32*|MSYS*) exe=$exe.exe ;;
 esac
 
-echo "Downloading $exe v$latest_tag"
+echo "Downloading v$latest_tag of $exe"
 tmp="$(mktemp)"
-curl -# --location --output "$tmp".sha256s.txt "$page/releases/download/$latest_tag/sha256s.txt"
-curl -# --location --output "$tmp"             "$page/releases/download/$latest_tag/$exe"
+curl -# --fail --location --output "$tmp"             "$page/releases/download/$latest_tag/$exe"
+echo Verifying checksum...
+curl -# --fail --location --output "$tmp".sha256s.txt "$page/releases/download/$latest_tag/sha256s.txt"
 tmpdir="$(dirname "$tmp")"
 ( cd "$tmpdir"
   mv "$tmp" "$exe"
@@ -60,9 +69,9 @@ tmpdir="$(dirname "$tmp")"
   rm "$tmp".sha256s.txt
   chmod +x "$exe"
 )
-mv -v "$tmpdir/$exe" "$target_path"/monkey
+mv -v "$tmpdir/$exe" "$target"
 
-installed_version=$("$target_path"/monkey --version)
+installed_version=$("$target" --version)
 if [ "$installed_version" != "$latest_version" ]; then
     fatal "This is not the expected version: $installed_version"
 fi
